Create Student database
This program has been made with help of python and file handling has been used for building it. This program helps us store names of students, remove a student name and/or search for names.  User defined functions has been also made to make the program look neat and clean and short too.